added posts endpoints

+++ server/routes/courses.js
@@ -23,18 +23,23 @@ courseRouter.get('/', async (req, res) => {
 });
 // Create a new course
 courseRouter.post('/',passport.authenticate("session") ,(req, res) => {
+    if(req.user){
     // check if user has role 2 or 3 (admin or teacher)
-    if (req.user.role === 1 || req.user.role === 2) {
-        Courses.create(req.body, (err, course) => {
-            console.log(req.body);
-            if (err) {
-                console.log(err);
-                res.status(500).send(err);
-            }
-            res.send(course);
-        });
-    } else {
-        res.status(401).send("You are not authorized to create a course");
+        if (req.user.role === 1 || req.user.role === 2) {
+            Courses.create(req.body, (err, course) => {
+                console.log(req.body);
+                if (err) {
+                    console.log(err);
+                    res.status(500).send(err);
+                }
+                res.send(course);
+            });
+        } else {
+            res.status(401).send("You are not authorized to create a course");
+        }
+    }
+    else{
+        res.sendStatus(401);
     }
 });
 
@@ -52,17 +57,26 @@ courseRouter.get('/:id', (req, res) => {
 
 // delete a course
 courseRouter.delete('/:id', passport.authenticate("session"), (req, res) => {
-    // check if user has role 2 or 3 (admin or teacher)
-    if (req.user.role === 1 || req.user.role === 2) {
-        Courses.findByIdAndRemove(req.params.id, (err, course) => {
-            if (err) {
-                console.log(err);
-                res.status(500).send(err);
-            }
-            res.send(course);
-        });
-    } else {
-        res.status(401).send("You are not authorized to delete a course");
+    if(req.user){
+        if (req.user.role === 1 || req.user.role === 2) {
+            Courses.findByIdAndRemove(req.params.id, (err, course) => {
+                if (err) {
+                    console.log(err);
+                    res.status(500).send(err);
+                }
+                if(course){
+                res.send(course);
+                }
+                else{
+                    res.sendStatus(404);
+                }
+            });
+        } else {
+            res.status(401).send("You are not authorized to delete a course");
+        }
+    }
+    else{
+        res.sendStatus(401);
     }
 });
